Move code into separate modules
Now logically grouped functions have a home of their own :)
